+++ docs/en/changes/changes.md
@@ -29,6 +29,7 @@
 * `BanyanDB Client`: Property query support `Order By`.
 * MQE: trim the label values condition for the labeled metrics query to enhance the readability.
 * PromQL service: fix time parse issue when using RFC3339 time format for querying.
+* Envoy metrics service receiver: support adapter listener metrics.
 
 #### UI
 * Fix the missing icon in new native trace view.